fix/move Login, Registration, ForgotPassword, Main, Profile to the pages directory
